requirements.txt
absl-py==0.12.0
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
configparser==5.0.2
docker-pycreds==0.4.0
gitdb==4.0.7
GitPython==3.1.17
google-auth==1.30.0
google-auth-oauthlib==0.4.4
grpcio==1.38.0
h5py==3.2.1
idna==2.10
joblib==1.0.1
Markdown==3.3.4
numpy==1.20.3
oauthlib==3.1.0
pandas==1.2.4
pathtools==0.1.2
Pillow==8.2.0
prettytable==2.1.0
promise==2.3
protobuf==3.17.0
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.3
sentry-sdk==1.1.0
shortuuid==1.0.1
six==1.16.0
smmap==4.0.0
subprocess32==3.5.4
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
threadpoolctl==2.1.0
torch==1.8.1
torchvision==0.9.1
typing-extensions==3.10.0.0
urllib3==1.26.4
wandb==0.10.30
wcwidth==0.2.5
Werkzeug==2.0.1
